Refactor project: use class, add methods
Redid the project to be used with class.

Got stuck on making projects dynamic with the use of objects and arrays,
but figured it out in the end. This solution caused further issues when
needing to push elements to a certain array inside the projects object.
This too was solved by looking up its index values inside the
allProjects array.

Also, cleaned up the code and tried to logically organise the modules.
It still does not feel too intuitive.